Move `index.js` to `resolver.js` and include `resolver.js` in the package.
