+++ install.js
@@ -1,4 +1,4 @@
-const PCR = require("./index.js");
+const PCR = require("./resolver.js");
 //console.log(PCR.getStats());
 PCR({}).then((pcr) => {
     //console.log(pcr);
