+++ package.json
@@ -2,7 +2,8 @@
   "name": "puppeteer-chromium-resolver",
   "version": "5.2.0",
   "description": "Tool to resolve puppeteer and chromium faster, detect local installed chromium, download chromium with custom mirror host, cache chromium revision out of node_modules, test chromium headless being launchable.",
-  "main": "index.js",
+  "main": "resolver.js",
+  "files": ["resolver.js"],
   "dependencies": {
     "gauge": "^3.0.0",
     "object-assign": "^4.1.1",
