+++ resolver.js
